initial commit as supertini
supertini runs any binary that was passed as the first argument of its
execution. It will then watch for changes in the directory the binary is
in. If the binary is removed, the current binary execution is stopped.
If a new binary appears, the binary is executed again. An arg file named
after the binary and suffixed with `.args` is expected too.

Signed-off-by: Sébastien Han <seb@redhat.com>